backend/requirements.txt
aiosqlite==0.20.0
fastapi==0.111.0
pyserial==3.5
SQLAlchemy==2.0.30
uvicorn==0.29.0
python-dotenv==1.0.1